feat: convert word/excel to pdf with soffice
