+++ backend/src/Program.cs
@@ -107,6 +107,7 @@ foreach (var module in modules)
 // Register global services
 builder.Services.AddScoped<ExcelTemplateService>();
 builder.Services.AddScoped<WordTemplateService>();
+builder.Services.AddScoped<PDFConverterService>();
 
 var app = builder.Build();
